Remove now dead code
