+++ lxd/api10.go
@@ -28,10 +28,6 @@ var api10 = []Command{
 	imagesCmd,
 	imagesExportCmd,
 	imagesSecretCmd,
-	operationsCmd,
-	operationCmd,
-	operationWait,
-	operationWebsocket,
 	newOperationsCmd,
 	newOperationCmd,
 	newOperationWait,
