+++ lxd/response.go
@@ -21,22 +21,36 @@ type resp struct {
 	Status     string            `json:"status"`
 	StatusCode shared.StatusCode `json:"status_code"`
 	Metadata   interface{}       `json:"metadata"`
+	Operation  string            `json:"operation"`
 }
 
 type Response interface {
 	Render(w http.ResponseWriter) error
 }
 
+// Sync response
 type syncResponse struct {
 	success  bool
 	metadata interface{}
 }
 
-/*
-  fname: name of the file without path
-  headers: any other headers that should be set in the response
-*/
+func (r *syncResponse) Render(w http.ResponseWriter) error {
+	status := shared.Success
+	if !r.success {
+		status = shared.Failure
+	}
 
+	resp := resp{Type: lxd.Sync, Status: status.String(), StatusCode: status, Metadata: r.metadata}
+	return WriteJSON(w, resp)
+}
+
+func SyncResponse(success bool, metadata interface{}) Response {
+	return &syncResponse{success, metadata}
+}
+
+var EmptySyncResponse = &syncResponse{true, make(map[string]interface{})}
+
+// File transfer response
 type fileResponseEntry struct {
 	identifier string
 	path       string
@@ -48,10 +62,6 @@ type fileResponse struct {
 	files            []fileResponseEntry
 	headers          map[string]string
 	removeAfterServe bool
-}
-
-func FileResponse(r *http.Request, files []fileResponseEntry, headers map[string]string, removeAfterServe bool) Response {
-	return &fileResponse{r, files, headers, removeAfterServe}
 }
 
 func (r *fileResponse) Render(w http.ResponseWriter) error {
@@ -118,108 +128,11 @@ func (r *fileResponse) Render(w http.ResponseWriter) error {
 	return err
 }
 
-func WriteJSON(w http.ResponseWriter, body interface{}) error {
-	var output io.Writer
-	var captured *bytes.Buffer
-
-	output = w
-	if *debug {
-		captured = &bytes.Buffer{}
-		output = io.MultiWriter(w, captured)
-	}
-
-	err := json.NewEncoder(output).Encode(body)
-
-	if captured != nil {
-		shared.DebugJson(captured)
-	}
-
-	return err
+func FileResponse(r *http.Request, files []fileResponseEntry, headers map[string]string, removeAfterServe bool) Response {
+	return &fileResponse{r, files, headers, removeAfterServe}
 }
 
-func (r *syncResponse) Render(w http.ResponseWriter) error {
-	status := shared.Success
-	if !r.success {
-		status = shared.Failure
-	}
-
-	resp := resp{Type: lxd.Sync, Status: status.String(), StatusCode: status, Metadata: r.metadata}
-	return WriteJSON(w, resp)
-}
-
-/*
- * This function and AsyncResponse are simply wrappers for the response so
- * users don't have to remember whether to use {}s or ()s when building
- * responses.
- */
-func SyncResponse(success bool, metadata interface{}) Response {
-	return &syncResponse{success, metadata}
-}
-
-var EmptySyncResponse = &syncResponse{true, make(map[string]interface{})}
-
-type async struct {
-	Type       lxd.ResponseType    `json:"type"`
-	Status     string              `json:"status"`
-	StatusCode shared.StatusCode   `json:"status_code"`
-	Operation  string              `json:"operation"`
-	Resources  map[string][]string `json:"resources"`
-	Metadata   interface{}         `json:"metadata"`
-}
-
-type asyncResponse struct {
-	run       func(id string) shared.OperationResult
-	cancel    func(id string) error
-	ws        shared.OperationWebsocket
-	resources map[string][]string
-	metadata  shared.Jmap
-	done      chan shared.OperationResult
-}
-
-func (r *asyncResponse) Render(w http.ResponseWriter) error {
-	op, err := createOperation(r.metadata, r.resources, r.run, r.cancel, r.ws)
-	if err != nil {
-		return err
-	}
-
-	err = startOperation(op)
-	if err != nil {
-		return err
-	}
-
-	body := async{Type: lxd.Async, Status: shared.OK.String(), StatusCode: shared.OK, Operation: op}
-	if r.ws != nil {
-		body.Metadata = r.ws.Metadata()
-	} else if r.metadata != nil {
-		body.Metadata = r.metadata
-	}
-
-	if r.resources != nil {
-		resources := make(map[string][]string)
-		for key, value := range r.resources {
-			var values []string
-			for _, c := range value {
-				values = append(values, fmt.Sprintf("/%s/%s/%s", shared.APIVersion, key, c))
-			}
-			resources[key] = values
-		}
-		body.Resources = resources
-	}
-
-	w.Header().Set("Location", op)
-	w.WriteHeader(202)
-
-	return WriteJSON(w, body)
-}
-
-func AsyncResponse(run func(id string) shared.OperationResult, cancel func(id string) error) Response {
-	return &asyncResponse{run: run, cancel: cancel}
-}
-
-func AsyncResponseWithWs(ws shared.OperationWebsocket, cancel func(id string) error) Response {
-	return &asyncResponse{run: ws.Do, cancel: cancel, ws: ws}
-}
-
+// Operation response
 type operationResponse struct {
 	op *newOperation
 }
@@ -235,7 +148,7 @@ func (r *operationResponse) Render(w http.ResponseWriter) error {
 		return err
 	}
 
-	body := async{
+	body := resp{
 		Type:       lxd.Async,
 		Status:     shared.OK.String(),
 		StatusCode: shared.OK,
@@ -252,12 +165,13 @@ func OperationResponse(op *newOperation) Response {
 	return &operationResponse{op}
 }
 
-type ErrorResponse struct {
+// Error response
+type errorResponse struct {
 	code int
 	msg  string
 }
 
-func (r *ErrorResponse) Render(w http.ResponseWriter) error {
+func (r *errorResponse) Render(w http.ResponseWriter) error {
 	var output io.Writer
 
 	buf := &bytes.Buffer{}
@@ -282,17 +196,17 @@ func (r *ErrorResponse) Render(w http.ResponseWriter) error {
 }
 
 /* Some standard responses */
-var NotImplemented = &ErrorResponse{http.StatusNotImplemented, "not implemented"}
-var NotFound = &ErrorResponse{http.StatusNotFound, "not found"}
-var Forbidden = &ErrorResponse{http.StatusForbidden, "not authorized"}
-var Conflict = &ErrorResponse{http.StatusConflict, "already exists"}
+var NotImplemented = &errorResponse{http.StatusNotImplemented, "not implemented"}
+var NotFound = &errorResponse{http.StatusNotFound, "not found"}
+var Forbidden = &errorResponse{http.StatusForbidden, "not authorized"}
+var Conflict = &errorResponse{http.StatusConflict, "already exists"}
 
 func BadRequest(err error) Response {
-	return &ErrorResponse{http.StatusBadRequest, err.Error()}
+	return &errorResponse{http.StatusBadRequest, err.Error()}
 }
 
 func InternalError(err error) Response {
-	return &ErrorResponse{http.StatusInternalServerError, err.Error()}
+	return &errorResponse{http.StatusInternalServerError, err.Error()}
 }
 
 /*
