+++ lxd/util.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"bytes"
+	"encoding/json"
+	"io"
+	"net/http"
+
+	"github.com/lxc/lxd/shared"
+)
+
+func WriteJSON(w http.ResponseWriter, body interface{}) error {
+	var output io.Writer
+	var captured *bytes.Buffer
+
+	output = w
+	if *debug {
+		captured = &bytes.Buffer{}
+		output = io.MultiWriter(w, captured)
+	}
+
+	err := json.NewEncoder(output).Encode(body)
+
+	if captured != nil {
+		shared.DebugJson(captured)
+	}
+
+	return err
+}
